Qukey for home row mods
Customizing Quekey plugin to work more smoothly
with home row mods. A test file is included for
quickly typing words with repeated home row keys
in the QWERTY layout.

The primary issues for me is quickly typing words
with repeated home row keys such as `less` and
`add`. The second `s` or `d` is often not
registered when Quekey plugin is used with
secondary modifier actions.

It's interesting that `less` is more problematic than
`add`. I *think* this is because `e` is on the top
row and has no seconary action configured. As opposed
to `add` where the `a` is also on the home row with
a seconary action configured.

To summarize the problem:

- When a non-Quekey modifier is quickly followed by
  two repeated tapped Quekey modifier keys, the second Quekey
  modifier key is incorrectly registered as holding the key
  to take on its alternate (modifier) state, instead of the
  intended primary tapped state.

- This happens less frequently when two repeated tapped Quekey
  modifer keys are followed by another Quekey modifier key.

- For example, with home row mods on QWERTY layout, the
  quickly tapped sequence `ess`, yields only one `s`.
  Presumably the second `s` is interpreted as a hold and
  takes on the modifier state.

(This needs further testing and confirmation.)